requirements.txt
elevenlabs==0.2.27
python-dotenv==1.0.1
python-telegram-bot==20.7
